+++ app/commands/homebrew_context/daggerheart/change_subclass_command.rb
@@ -13,13 +13,14 @@ module HomebrewContext
         params do
           required(:subclass).filled(type?: ::Daggerheart::Homebrew::Subclass)
           optional(:name).filled(:string, max_size?: 50)
+          optional(:public).filled(:bool)
         end
       end
 
       private
 
       def do_persist(input)
-        input[:subclass].update!(input.slice(:name))
+        input[:subclass].update!(input.slice(:name, :public))
 
         refresh_user_data.call(user: input[:subclass].user) if input[:subclass].user