+++ app/javascript/applications/HomebrewsApp/assets/index.jsx
@@ -1,4 +1,5 @@
 export * from './Chevron';
+export * from './Copy';
 export * from './Edit';
 export * from './Plus';
 export * from './Stroke';